refactor(FileList): move FileListItem to separate file
