+++ packages/picasso/src/FileList/test.tsx
@@ -1,7 +1,11 @@
 import React from 'react'
-import { render, fireEvent } from '@toptal/picasso/test-utils'
+import { render } from '@toptal/picasso/test-utils'
+import { OmitInternalProps } from '@toptal/picasso-shared'
 
-import FileList from './'
+import FileList, { Props } from './FileList'
+
+const renderFileList = (props: OmitInternalProps<Props>) =>
+  render(<FileList {...props} />)
 
 describe('FileList', () => {
   const file = {
@@ -12,78 +16,8 @@ describe('FileList', () => {
   }
 
   it('renders', () => {
-    const { container } = render(<FileList files={[file]} />)
+    const { container } = renderFileList({ files: [file] })
 
     expect(container).toMatchSnapshot()
-  })
-
-  describe(`when 'onItemRemove' is provided`, () => {
-    it('renders remove button', () => {
-      const handleIItemRemove = jest.fn()
-
-      const { getByRole } = render(
-        <FileList files={[file]} onItemRemove={handleIItemRemove} />
-      )
-
-      fireEvent.click(getByRole('button'))
-
-      expect(handleIItemRemove).toHaveBeenCalled()
-    })
-  })
-
-  describe('when file is uploading', () => {
-    it(`renders 'Uploading...' label and progress bar`, () => {
-      const { queryByText, queryByTestId } = render(
-        <FileList files={[{ ...file, uploading: true, progress: 30 }]} />
-      )
-
-      expect(queryByText('Uploading...')).toBeInTheDocument()
-      expect(queryByTestId('file-list-item-progressbar')).toBeInTheDocument()
-    })
-
-    describe('when error exists', () => {
-      it('renders file name and error message', () => {
-        const { queryByText, queryByTestId } = render(
-          <FileList
-            files={[
-              {
-                ...file,
-                uploading: true,
-                progress: 30,
-                error: 'File is too large'
-              }
-            ]}
-          />
-        )
-
-        expect(queryByText('Uploading...')).not.toBeInTheDocument()
-        expect(
-          queryByTestId('file-list-item-progressbar')
-        ).not.toBeInTheDocument()
-
-        expect(queryByText('user-profile-picture.png')).toBeInTheDocument()
-        expect(queryByText('File is too large')).toBeInTheDocument()
-      })
-    })
-  })
-
-  describe('when error exists', () => {
-    it('renders file name and error message', () => {
-      const { queryByText } = render(
-        <FileList
-          files={[
-            {
-              ...file,
-              uploading: true,
-              progress: 30,
-              error: 'File is too large'
-            }
-          ]}
-        />
-      )
-
-      expect(queryByText('user-profile-picture.png')).toBeInTheDocument()
-      expect(queryByText('File is too large')).toBeInTheDocument()
-    })
   })
 })
