+++ packages/picasso/src/FileList/types.ts
@@ -1,0 +1,6 @@
+export interface File {
+  uploading: boolean
+  progress?: number
+  error?: string
+  name: string
+}
