+++ packages/picasso/src/FileListItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FileListItem'
