+++ packages/picasso/src/FileListItem/styles.ts
@@ -1,0 +1,11 @@
+import { Theme, createStyles } from '@material-ui/core/styles'
+import { rem } from '@toptal/picasso-shared'
+
+export default ({ palette }: Theme) =>
+  createStyles({
+    root: {
+      borderBottom: `1px solid ${palette.grey.light2}`,
+      paddingTop: rem('4px'),
+      paddingBottom: rem('4px')
+    }
+  })
